Undo FakeLanguageManager move + Remove Translatable.languageManager() method
