+++ src/test/java/nova/wrappertests/NovaLauncherTestFactory.java
@@ -24,12 +24,12 @@ import nova.core.block.Block;
 import nova.core.block.BlockFactory;
 import nova.internal.core.Game;
 import nova.internal.core.bootstrap.DependencyInjectionEntryPoint;
-import nova.internal.core.depmodules.FakeLanguageModule;
 import nova.internal.core.launch.NovaLauncher;
 import nova.wrappertests.depmodules.FakeClientModule;
 import nova.wrappertests.depmodules.FakeComponentModule;
 import nova.wrappertests.depmodules.FakeGameInfoModule;
 import nova.wrappertests.depmodules.FakeKeyModule;
+import nova.wrappertests.depmodules.FakeLanguageModule;
 import nova.wrappertests.depmodules.FakeNetworkModule;
 import nova.wrappertests.depmodules.FakeRenderModule;
 import nova.wrappertests.depmodules.FakeSaveModule;
